skidl tools init
This adds a "Subcircuit" wrapper which can be used to expose the pins
of multicomponent subcircuits. It attempts to compensate for
deficiencies in skidl's Part object, which only supports single-unit
parts. Like Parts, Subcircuits can use dictionary-like getters and
setters. The intention behind this is to be able to interface with
these objects in an identical fashion.